Report all DongTai Thread Count.
