+++ dongtai-agent/src/main/java/io/dongtai/iast/agent/monitor/impl/ServerConfigMonitor.java
@@ -1,13 +1,16 @@
-package io.dongtai.iast.agent.monitor;
+package io.dongtai.iast.agent.monitor.impl;
 
 import io.dongtai.iast.agent.Constant;
+import io.dongtai.iast.agent.monitor.IMonitor;
+import io.dongtai.iast.agent.monitor.MonitorDaemonThread;
 import io.dongtai.iast.agent.report.AgentRegisterReport;
 import io.dongtai.iast.agent.util.http.HttpClientUtils;
 import io.dongtai.log.DongTaiLog;
 import org.json.JSONObject;
 
-public class ServerConfigMonitor implements IMonitor{
+public class ServerConfigMonitor implements IMonitor {
     private final String name = "ServerConfigMonitor";
+
     public JSONObject serverConfig = new JSONObject();
 
     @Override
@@ -17,6 +20,7 @@ public class ServerConfigMonitor implements IMonitor{
 
     @Override
     public void check() {
+        DongTaiLog.info(("Server Config Check"));
         getConfig();
     }
 
@@ -33,9 +37,9 @@ public class ServerConfigMonitor implements IMonitor{
      * 根据agentID获取服务端对Agent的配置
      */
     public void getConfig(){
+        DongTaiLog.info("Get Server Config");
         JSONObject report = new JSONObject();
         StringBuilder response;
-
         report.put(Constant.KEY_AGENT_ID, AgentRegisterReport.getAgentFlag());
         try {
             response = HttpClientUtils.sendPost(Constant.API_Server_Config,report.toString());
