+++ dongtai-agent/src/test/java/com/secnium/iast/agent/monitor/PerformanceMoniterTest.java
@@ -1,7 +1,7 @@
 package com.secnium.iast.agent.monitor;
 
 import io.dongtai.iast.agent.manager.EngineManager;
-import io.dongtai.iast.agent.monitor.PerformanceMonitor;
+import io.dongtai.iast.agent.monitor.impl.PerformanceMonitor;
 import org.junit.Test;
 
 import java.lang.management.ManagementFactory;
